Stop hard-coding the framework path
It may happen a user doesn't want to copy the framework in the
"standard" location, or that they are using some other tools that clone
the framework from another provider.

In such case, controller-0 might see the framework cloned in some other
tree, meaning all of the generated content.

Let's try to be more flexible by using the user provided data, while
defaulting to the standard location.

The internal _cifmw_reproducer_framework_location fact is available
since the very start of the reproducer role call, ensuring we're able to
use it even for new features.