Modify libvirt_manager to run linear instead of batches
The initial approach was batching VMs, meaning we were having multiple
times the same kind of `wait for SSH` and related "pause".

This patch changes how we define, start and configure the various VM, by
using `cifmw_libvirt_manager_all_vms` available since #1777.
This list is showing all of the VM created by the libvirt_manager role,
and with "simple" string transformation, we can extract the VM family
(compute, ocp, networker, etc) so that we can still access the various
configurations for each of them.

This linear approach allows to leverage `async` in various places, as
well as some other helpers to make the whole run smoother and faster.