Switch to dhcp-hostsfile
We hit a race condition with ansible creating the DHCP file with strict
rights, ionitofy kicking in, dnsmasq trying to read the file and
failing.
This prevented the service to load a DHCP host configuration, leading to
issues.

With this change, we ensure we keep a directory, but we have to trigger
the reload ourself - like before. This ensures the rights are all
applied.