Enable forward in libvirt zone
For ironic usecase we need to allow traffic forwarding between networks
in libvirt - without nat.

Add option:
  cifmw_libvirt_manager_firewalld_zone_libvirt_forward

Defaults to `true`, this is the behaviour we tried(?) to have with the
code in devscripts that this PR removes.

Also, by default set masquerade on the "default" zone i.e public. This
allows us to remove the iptables rules inserted by the devscripts role.