Ensure we rely only on dnsmasq
This patch corrects and finalize the move from CRC hard-coded hosts
entries to full dnsmasq DHCP and DNS.

In a previous PR (#1908) we switched CRC dns entries to dnsmasq, but the
base CRC config weren't cleaned, leading to a broken state.
It was mostly working because controller-0 still had some hardcoded
entries in /etc/hosts.

This patch fixes the initial issue with dnsmasq configuration overlaps
(ours vs CRC), and ensures we don't rely on /etc/hosts entries on
controller-0.