[DUP without nodeset]Add controller-0 group name in zuul nodeset
In crc reproducer, we have controller-0 as a host for ansible controller
but in Zuul CI, we use controller as a hostname.

When fetch_compute_facts runs on controller-0 in CI, it does not resolve
the hostname in CI but should work in reproducer.

In order to fix that, we are adding a controller-0 group in the zuul
extracted crc nodeset to fix that.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>