Avoid printing secret with CA in install_openstack_ca
When we get the combined-ca-bundle secret to install the CA on the
controller, it prints the content of the secret to stdout. This content
is ofter really big, making the logs much longer without adding much
benefit, so this change adds no_log: true to avoid these lines in the
logs.