nat64 - pin dnsmasq to v2.89 tag
Experiencing some issues, the nat64 stops responding to dns requests and
strace on unbound shows:
  recvfrom(3, 0x5641a194bcf0, 65552, 0, 0x7ffcbbc34438, [128]) = -1 EAGAIN (Resource temporarily unavailable)

I guess the main branch of dnsmasq is having some issue. Let's try to
pin on the v2.89 release tag.

Also add retries to the ping test in the molecule coverage. The same
test fails on PR#1976.