nat64 - Fix molecule, and touch ups to stabilize
The molecule tests was failing because the interfaces did not end up in
the 'libvirt' firewall zone.

Added the zone attribute to the libvirt bridge element. This should work
according to libvirt documentation, but for some reason not in this
test environment. So, had to do that manually with firewalld.

Also:
 * Pin dnsmasq to the v2.89 tag.
   Should be more stable than running of the main branch with frequent changes.
 * Set do-daemonize: no in unbound config.
   This removes a warning logged by the unbound service, do-daemonize and
   use-systemd should not be used togheter.

There is a lot of "crude" debug info gathering added in the molecule
tests, I think we should keep that around as is to make troubleshooting
any potentital future issues in this job.