Add a hook to increase quotas for project service
When running Glance using Cinder as a backend the `service` project is
being used and tests can go over the volume quota.

This patch adds a playbook hook to increase the volume quota for the
`service` project to 50.

Uses the region set in the `.config/openstack/clouds.yaml` of the
`openstackclient` pod, and assumes the project name is `service`.