dnsmasq: Use network/prefix in --domain option
We add the `,local` option to --domain configuration statements.
According to dnsmasq manual page this only works if using network/prefix
for the range.

```
--domain=<domain>[[,<address range>[,local]]|<interface>]

If the address range is given as ip-address/network-size, then a additional
flag "local" may be supplied which has the effect of adding --local declarations
for forward and reverse DNS  queries.
Eg.   --domain=thekelleys.org.uk,192.168.0.0/24,local is identical to
      --domain=thekelleys.org.uk,192.168.0.0/24 --local=/thekelleys.org.uk/ --local=/0.168.192.in-addr.arpa/
```