Add variables to check if the pod is running or not
Usually, the default values to check if the pod is running or not is
enough for normal jobs, but we notice an increase in time waiting when
we are running unijobs, maybe because the amount of services running.
Adding variables to set this values, will ensure we can defined a bigger
timeout while checking for this without disrupt other jobs.