Correct prefix replacement in network names
Usually we have `cifmw-` prefix, but the `replace` was looking for
`cifmw_` instead.
While it didn't create any issue in CI nor apparently in most of the
other scenarios, DCN had an issue.

This patch replaces the `replace` by `regex_replace` to ensure we're
properly catching both of `cifmw_` and `cifmw-` like some other places.