Change iscsid login timeout in computes
In a previous patch we changed the iSCSI login timeout for the control
plane (OpenShift nodes).

In this patch we add a new hook to do the same thing for the edpm nodes,
because the default timeout of 2 minutes is too high for some test
scenarios.

It is necessary to do it this way because edpm-ansible doesn't currently
have a mechanism to change the iscsid.conf file. Once that feature is
added this patch can be reverted.

This patch changes the edpm iscsid timeout default to 3 retries and 5
seconds each (15 seconds in total), which is convenient for testing, as
any healthy deployment and backend should be able to login to the
backend in that amount of time, and if there is a broken path it will
not take 2 minutes to give up, just around 15 seconds.

Related-Jira: https://issues.redhat.com/browse/OSPRH-7614