Do not gather facts from networking_mapper on controller-0
Since we provide the interfaces info, we don't really need to connect to
the hosts to gather the networking facts: we already have everything we
need.

This also means faster infra deployment, since the couple of tasks
associated to the remote facts takes over 2 minutes.

It also converges some paths to use an existing parameter instead of
re-writing over and over `/home/zuul/ci-framework-data`.