Setup minimum environment to run update on architecture.
When the deployment is not using `install_yaml` we need to set up some
variables to be able to trigger the update.

The update is using install_yamls in two places. First, it uses the
`set_openstack_containers` to modify the `openstack-operator` and that
role leverage the `install_yamls` commands.

Second, it uses it for the update itself as it's a target of
`install_yamls`