Prepare crc job for architecture repo
Create a job that can deploy the HCI VA with an extracted crc layout.
This change introduces the jobs definition and playbooks for such a job.
It creates and extracted crc layout using ci-bootstrap role
and then calls networking mapper to get the required networking config to
deploy a VA. It needs to generate an interfaces-info file.

Additionally, it changes the 06-deploy-architecture playbook to allow
ssh access from any network to the compute nodes. It also modifies the
multinode-customizations playbook to work with a 'ctlplane' network in
the crc, instead of a 'default' one.

It makes use of scenario inhertance proposed in [1] to reuse most of the
existing HCI VA scenario.

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/1491

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/1491