Drop all references to change_message from zuul vars
After a new update, the change_message is appearing in more fields. As
observed recently [1], if the first comment on the PR contains some mock
ansible, if that is saved into the zuul_vars, it'll crash when trying to
load them. This change drops the newly introduced vars 'build_refs' and
'buildset_refs', whihc contain the change_message.

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/1879