OSPCIX-372 Fix race in ironic tls patch sequence
In order for ironic agent ramdisks to work, they need to contain the
CA certificate of the deployment such that they can talk to the OCP
TLS endpoint.

Unfortunately, the sequence of invocation to patch the ramdisk
images, which was merged, contains an invocation of a podman
plugin, which by default launches containers as detached
operations. The net result is the command can return
almost immediately, and then the next step attempts to
execute and then fails because files it expects may not appear
for several seconds. So the correct path is to disable
the detach operation call so when podman exits, we know the task
has been completed to allow for the next task to execute.

Also adds an ls -la of the folder we expect files to be in to aid
in future ci framework troubleshooting.