Added cifmw_install_ca_urls var to pass multiple ca urls
Currently cifmw_install_ca_url takes a single url. What if an user
wants to pass multiple ca urls then cifmw_install_ca_urls will help us
to achieve that.

Since cifmw_install_ca_url is used already. So we are not removing it
to avoid breakage.

It is just an improvement to the role to pass multiple ca urls.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>