Improve async status output for Zuul
It seems Zuul is buffering the async poll status, meaning the "hanging"
task shown in the job console is the one just before the actual deploy,
"Copy the configuration file."
It's slightly misleading and wrong.

This patch decouple the poll from the install task, and should therefore
display the right task in the console, even with the buffering.
The retries is computed based on the installation timeout.