Do not edit sensitive data set in machineconfig
By editing the sensitive data for the OpenShift, later the machineconfig
daemon will set the node as degraded, so later the operations that
we would like to be done via machineconfig manifests, will be not
triggered.
Some projects are using the machineconfig manifests to configure the
host, not directly by using the Ansible.