Rework a bit rhol_crc
There were multiple issues in the role:
- no version check, meaning it would consume whatever version it finds
  on the disk instead of ensuring it's consuming the target version
- lack of molecule tests for some specific and sensitive parts

This change also moves the "crc" binary to a more conventional location
(~/bin) to allow an easier discovery, and avoiding potential conflicts.

It also removes some wrong/useless parameters that could lead to weird
situations regarding rights, users and so on.

This PR also bumps the CRC version in both role and molecule (this
creates a conflict with another PR, #2102).