Add a tag to the "fix marketplace" imported tasks
Adding this tag will allow people to get a better re-run experience:
when running multiple time the role, it may happen the marketplace is
broken "for good".

This tag allows now to get a first run up to whatever stage you want,
the re-run with `--skip-tags fix_marketplace` to avoid "unfixing" it.