Correct generated OCP domain
A mistake was introduced when we generate the `cifmw_reproducer_domain`:
the ocp_cluster was missing, leading to "some" issues with the DNS
resolution whenever we wanted to use that domain:
dnsmasq is configured to delegate cluster_name.base_domain to our own
instance, so the "compute-0.base_domain" couldn't resolv.

This patch corrects the issue, allowing to ease the move to FQDN on the
various machines.

Note that, for now, nothing in the Framework is using that specific
domain - all of the accesses are done using the `.utility` domain.