Ensures we can use FQDN in ceph.yml playbook
Until now, there was an inconsistency between the 06-deploy-edpm.yml and
06-deploy-architecture.yml playbooks when it comes to ceph
configuration:
in the first one, ceph.yml play was instructed to use FQDN (via
ansible_fqdn fact), while the latter was using shortnames (via
ansible_hostname).

This leads to issues whenever we want to set FQDN on the various nodes
and in the inventory, since some stages of the ceph configuration is
checking for name consistency.