Expose found versions and location for crc binary
This will make things slightly easier to understand when the role is
wanting to install a new CRC binary/bundle.

It also change a task title to reflect the new behavior, and changes an
import_tasks to include_tasks since there's a `when` condition, to make
it slightly faster/better.