Use inventory_hostname in ceph related roles and play
Until now, the ceph playbook and associated roles were trying to be
smart, but it created many other issues, and continues to create weird
situations.

With this patch, we reverse the burden, and base everything on the
inventory.
Since the inventory is generated by the Framework, and the Framework
also sets the virtual machines hostname, we should be able to rely on
the `inventory_hostname` only.

This means the inventory must be consistent, if provided by external
means.

We therefore update/correct the hostname set on the various nodes to
match the inventory.