[test-operator] Fix volume mounting
In this patch we introduced a small bug that appears when we run both
tempest and tobiko tests in a single job [1]. The issue is that when
tobiko is executed both variables volume and volume_mounts contain
values from the tempest run.

This PR ensures that both volume and volume_mounts are initialized to
an empty list when logs are being collected. Additionaly, this patch
fixes the copying of logs from logs pod.

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/2124