Use cifmw_dlrn_report_promote_source to set proper dlrn reporting target
In update job, we deploy the system using cifmw_repo_setup_promotion: current-podified
var and update it to podified-ci-testing via different update vars.

dlrn_report role works on repo-setup vars. In this case, dlrn will
report the job status against current-podified. But we want to report
the status against podified-ci-testing. In this case, current
dlrn-report does not work.

In order to fix that we are adding a new var `cifmw_dlrn_report_promote_source` to set
the proper dlrn target against which job should report. It will not
affect existing job. As it's value depends on cifmw_repo_setup_promotion
case.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>