Consolidate conditions triggering re-deploy
Until now, the role didn't properly detect the need to re-deploy the OCP
cluster, leading to some bad situations.

This patch corrects that situation, and improve the "cleanup.yml" in
order to ensure it also removes any dangling base image - that might be
the case when a fist deploy involved workers, but the next one doesn't:
the dev-scripts configuration wouldn't know about it, since it gets
overridden, and we'd be left with the worker images dangling. Such a
situation would lead in the next runS to re-deploy OCP every single time
if we don't include workers anymore.