Added cifmw_update_containers_prefix var to set container prefix
In Upstream the openstack-services container name starts with openstack.
In downstream released content, the container name starts with
rhoso-openstack.

In this case, update_containers role will not work. As container names
are different.

Adding cifmw_update_containers_prefix var makes it easier to achieve the
same.

This pr also constructs the container url in a temprory var and reuse it
across the template to make it more readable.

Signed-off-by: Chandan Kumar (raukadah) <raukadah@gmail.com>