Ensure we get a timeout for node listing
It seems getting the nodes might hang, and it never ever returns.

This patch ensures we set a wait_condition, meaning we can toggle the
`wait` parameter to `true` and benefit from the default `wait_timeout`
of 120s[1] - which should be more than enough to get the OCP cluster
nodes list.

This PR also extract the node listing into its own tasks file, so that
we can import it instead of copy/pasting the same code.

This PR also adds a debugging capability when we're waiting for the
cluster to stabilize after removing the certificates. Using our
ci_script action, we're able to get logs, meaning we can easily review
the command output. Usually it lists the failing service(s) and provides
data about stability duration.

[1] https://docs.ansible.com/ansible/latest/collections/kubernetes/core/k8s_info_module.html#parameter-wait_timeout