Enable unique names for certain types of VMs
Unique names are needed for some specific usecases, but not for all VMs.

This patch ensures anything that's not OCP/CRC nor controller VM gets a
unique, random string in their hostname.

We also converge names, ensuring no underscore (_) are present in the VM
names - we replace them with dashes (-) to ensure it's a valid hostname.
This leads to a modification in the way we build the VM listing, it's
now a dict associating the name to the type, in order to make things
"slightly" easier to handle later.

This patch introduces a new top-level parameter, `cifmw_run_id`, that
can be either passed down by the user, or generated using the
random_string lookup from the community.general collection.

Allowing to set the custom ID would allow, for instance, to use the Zuul
job ID as the identifier, making things slightly easier to follow
across the various runs.