Configure Ceph RGW on a dedicated Swift network
When Ceph RGW is used, an endpoint for Swift storage is
hosted not in a pod on k8s but on an EDPM node. Thus, a
service hosted on an EDPM node will need to be accessed
from a separate network.

This patch adds the Swift storage network (swift) with
VLAN 25 and range 172.22.0.0/24 in the reproducers
networking-definition.yml. It also adds a multus range
for this network so that the Tempest pod can access this
network for testing. The Swift storage network is added
to the OCP nodes for the same reason. The swift network
is also added to the ci playbook networking-env-definition
for the rhoso-architecture-validate-hci github-check.

The Ceph playbook cifmw_cephadm_rgw_network parameter may
then be set to the new swift network so that RGW is then
configured on a separate network. If this network
parameter is not set, then the storage network is used.

Signed-off-by: John Fulton <fulton@redhat.com>