Ensure we don't have shortnames in the DNS
Without the dhcp-fqdn option, dnsmasq might push shortnames in the DNS,
leading to some weird situation with the resolution of such names.

Before that change, "dig compute-0" might resolve to the public/ocpbm
network, or, later, on the osp_trunk/ctlplane network.

This is because, in the DHCP configuration, we use only shortnames,
meaning we get "compute-0" twice, with different IP/networks.
Then, it depends on the order of the resolvers in /etc/resolv.conf -
order that might change upon DHCP lease updates.

With this patch, we ensure no shortnames are injected, leading to a more
stable DNS resolution, while it can also lead to slower resolution,
since the request might have to go through multiple resolvers to get an
answer.