Optimize Ceph log collection and optionally dump to file
Instead of running several commands to get the required ceph logs, this
patch reworks the way the logs tasks are executed. It's now based on a
struct that can be iterated.
In addition, it might be a good idea to collect several logs, at
deployment time, in files, other then (optionally) printing them in the
playbook log file.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>