Introduce Ceph Upgrade tasks
This patch improves the existing cephadm role and introduces a set of
tasks that are supposed to perform a ceph update/upgrade.
The main idea behind this work is to be able to include it from [1]
and rollout new Ceph daemons before starting an OpenStack minor update.

[1] https://github.com/openstack-k8s-operators/ci-framework/blob/main/update-edpm.yml#L18

Jira: https://issues.redhat.com/browse/OSPRH-9697

Signed-off-by: Francesco Pantano <fpantano@redhat.com>