Allow to know what data were used
in hci_prepare, some assertions are made, to ensure we get the needed
data. Until now, it was failing, but we didn't know what data were
consumed.

This PR leverage the block/rescue pattern to output the checked data.

It also correct how those data were checked and consumed: the network
names are usually converged to a simple string without any dashes, at
least in a job reproducer case.
This patch allows to cover both names, the one from the CI job
definition (with dashes) and the one rebuilt over name conversions.