[polarion] Each result dir equals polarion test run
Each of the dirs found under the cifmw_polarion_jump_result_dir
will be understood as a new test run in polarion. The polarion
role will merge all .xml result files found inside each of the
directories. Then it will iterate over all merged files and will
upload them to polarion as a test run.

The reason behind this is that the same test can be executed
several times within one job, each time with a different
configuration or in a different cloud state (e.g. before/after
service reboot). However, if polarion receives multiple results
of the same test it will take into account only the last one.

The results from the tests that are executed repeatedly need to
be in separate result directories so that they are reported as
separate test runs.