Add a role to run compliance scans against the compute nodes and control plane (#2228)
* Add compliance role

* Add code to run scans on compute nodes

* Add playbook to run compliance scans

* Fix pre-commit errors

* Add molecule test files

* Add login credentials for registry.redhat.io

* Fix various errors

* Add check to ensure compliance-operator is ready

Also make sure that the scans has started and are complete.
And set to ignore errors on compute compliance scans.

* Fix the target directory for compute scan results

* Modify the script to pass compute scans correctly

* Avoid using shell when unzipping files

We make a couple modifications to the scan logic to avoid the use of
shell and add a zuul job for the molecule tests.  Also added some
changes to variables to make them more specific.

* Fix issues from code review