[Reproducer] Add support for disabling ocp overlays
This patch adds a new top level variable `cifmw_use_ocp_overlay` used
to control whether the OCP VMs created by dev_scripts via the reproducer
role use image overlays or not.

Prior to this patch the use of image overlays was not configurable, while image overlays are useful for local deployments, in CI they are
not used. The downside to image overlays is the process to create them
takes a very long time.

The OCP cluster needs come up and stabilize, then we shut the VMs down and copy the images to use as base images, we then need to start the
OCP VMs again and wait for it to stabilize, this is all wasted time
because in CI environments we start fresh every time.

This patch also removes the undocumented internal variable
`cifmw_use_reproducer` and it's use replaced with
`cifmw_use_ocp_overlay`

Jira: https://issues.redhat.com/browse/OSPRH-7771