Add role to deploy 17.1 env for adoption
Add a new role that will deploy a tripleo environment that will serve as
source for adoption. This role is expected to cosume the infra created
by [1], and a 17.1 scenario definition from the data-plane-adoption
repo, introduced by [2].

It also introduce a small fix to the deploy-ocp.yml so the resulting ocp
cluster is ready (the nodes needed to be uncordoned).

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/2285
[2] https://github.com/openstack-k8s-operators/data-plane-adoption/pull/597