Introduce two new playbooks for adoption infra
create-infra.yml allows to create all the VMs related to adoption
scenarios.
deploy-ocp.yml allows to get the OCP cluster in, independently to the
rest of the Framework. You usually want to run that playbook after the
"create-infra.yml", but it's not mandatory.

The infrastructure and networks changes needed for adoption are to be
pushed in a scenario file, either under scenarios/adoption/SCENARIO.yml,
or in a 3rd party repository (mostly for Downstream).

All of the VMs are shut off after the create-infra.yml is finished, and
only the OCP cluster members are running after the deploy-ocp.yml run.