Ensure we can re-run the framework without issue on config_drive
Until now, if we were to re-run the same deploy without any cleanup, it
would fail on the geniso from config_drive role because it can't
override the existing ISO.

Adding the "creates" parameter to the ci_script corrects this issue.
The assertions also ensures we don't try to edit an existing ISO - this
wasn't supported before, but since the run crashed, the user would know
there was an issue. The assertion makes it clear now.

The next issue is related to the usage of the ISO from within
libvirt_manager: the ISO isn't removed once it's attached, and
re-running fails since it's already attached to the VM.

Checking for the attached volumes, and removing the useless "removes"
from the command (converted to shell) ensures it's working fine.

This patch also attempts to fix the molecule jobs for libvirt_manager:
from time to time, it might crash due to the incomplet cleanup done in
the common "prepare" stage: the ocp_volumes pool is still present, but
the underneath directory is absent, leading to a crash whenever we want
to refresh the pools:
```
msg: 'cannot open directory ''/opt/basedir/ocp_volumes'': No such file or directory'
```