Set default value for cifmw_architecture_automation_file
After a recent change [1] we define the architecture automation to be
used in the reproducer, based on the VA used. The HCI VA crc job does
not use the reproducer role, so it does not get that parameter defined.

Set a default value for cifmw_architecture_automation_file in the
deploy-architecture playbook for the cases where the reproducer role was
not called before and the var is undefined.

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/2292